% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetBearing.r
\name{GetBearing}
\alias{GetBearing}
\title{Calculate bearing between points}
\usage{
GetBearing(x, y)
}
\arguments{
\item{x}{integer vector of x coordinates}

\item{y}{integer vector of y coordinates}
}
\value{
vector of bearings (the first one being NA)
}
\description{
Return the bearing between a sequence of points
}

